% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_r_functions_wrapping_commandline_tools.R
\name{cli_ask_default_based_on_type}
\alias{cli_ask_default_based_on_type}
\title{Ask for defualt based on type}
\usage{
cli_ask_default_based_on_type(type = c("unknown", "flag", "arg", "pos"))
}
\arguments{
\item{type}{type of argument}
}
\value{
default value
}
\description{
Ask for defualt based on type
}
