% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_r_functions_wrapping_commandline_tools.R
\name{cli_param_list_to_function}
\alias{cli_param_list_to_function}
\title{Title}
\usage{
cli_param_list_to_function(
  cli_arg_list,
  base_command,
  function_name,
  tool_title,
  tool_description = ""
)
}
\arguments{
\item{cli_arg_list}{cli_arg_list object}

\item{base_command}{name of base command (e.g. grep) (string)}

\item{function_name}{name of the R function that will wrap the commandline tool (string)}

\item{tool_title}{name of the tool we are wrapping. Used only roxygen documentation (string)}

\item{tool_description}{what does the cli tool we're wrapping actually do (string)}
}
\value{
run for its side effects (writes a simple CLI tool wrapping R function to the clipboard)
}
\description{
Title
}
\examples{
\dontrun{
# Creates a function that wraps a commandline tool and writes it to the clipboard
list() |>
  cli_add_arg(name = "threads", prefix = "--threads") |>
  cli_add_flag(name = "d4", prefix = "--d4", default = FALSE) |>
  cli_add_pos(name = "prefix", required = TRUE) |>
  cli_add_pos(name = "bam_or_cram", required = TRUE) |>
  cli_param_list_to_function(base_command = "mosdepth", function_name = "run_mosdepth", tool_title = "Mosdepth", tool_description = "simple R wrapper")
}
}
