% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen2cli.R
\name{is_param_flag_or_arg}
\alias{is_param_flag_or_arg}
\title{Is paramater a flag / arg based on param_description}
\usage{
is_param_flag_or_arg(param_description)
}
\arguments{
\item{param_description}{paramater description (character)}
}
\value{
"flag" for flag arguments and 'arg' for non-flag args (character)
}
\description{
If paramater description contains '(flag)' or '(boolean)' or '(bool)' assume argument is a flag
otherwise assume is an argument
}
