% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen2cli.R
\name{is_param_required}
\alias{is_param_required}
\title{Is paramater required}
\usage{
is_param_required(param_description)
}
\arguments{
\item{param_description}{paramater description (character)}
}
\value{
TRUE if paramater description contains '(required)', false if not
}
\description{
Is paramater required
}
