% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_r_functions_wrapping_commandline_tools.R
\name{readline_choose_yes_or_no}
\alias{readline_choose_yes_or_no}
\title{Ask yes or no}
\usage{
readline_choose_yes_or_no(prompt)
}
\arguments{
\item{prompt}{prompt for yes/no question (string)}
}
\value{
TRUE/FALSE depending on user response (Yes/No) (boolean)
}
\description{
Drop in replacement for askYesNo that will ask the question over and over until a yes/no answer is given
}
\examples{
\dontrun{
readline_choose_yes_or_no("Do you like apples?")
}
}
