% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen2cli.R
\name{test_function}
\alias{test_function}
\title{My Test Function}
\usage{
test_function(path, verbose, one_of = c("hi", "hello", "hiya"))
}
\arguments{
\item{path}{path argument}

\item{verbose}{should we print messages? (flag)}

\item{one_of}{choice of hi, hello, or hiya (required)}
}
\value{
some boolean
}
\description{
This is a simple oneline description
}
